refactor(server): extract parsers and app factory for testing
